Issue 831:  Add JSL into joblogs These changes are from review comments to message explanations and moving the formatXML function closer to the JSL serialization code.
